pyproject.toml
[project]
name = "information-retrieval"
version = "0.1.0"
description = "Repository for Information Retrieval homeworks"
requires-python = ">=3.14"
dependencies = [
    "pystemmer>=3.0.0",
]
